Start of Authentication Studies
After I built <my-login> I had an accident with Bower: I tried to
<paper-button> which overrode most Polymer2 base elements with Polymer1
and it ruined practically everything.
For next time- make sure you only import up to date components and that
their dependancies are also up to date.
The PowerShell is interactive so I use that to select when conflicts in
Bower arise.

Other than the authentication start I designed the app a bit more, I'll
deploy when auth is properly set up and running.
Still left to solve :
1. Why are the buttons invisible?
2.Why don't they call their ontap functions?
3.<dom-if> hiding - it worked fine with the standard <button> element
but not with <paper-icon-button>